pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-pm"
version = "0.1.0"
description = "Spec-driven autonomous PM and UX issue generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Raj Contractor"}]
dependencies = [
    "pydantic>=2.8.0,<3",
    "PyYAML>=6.0,<7",
    "playwright>=1.52.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9",
    "ruff>=0.11.0,<1",
    "mypy>=1.10.0,<2",
]

[project.scripts]
pm-agent = "pm_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -ra --strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
ignore_missing_imports = true
